feat: port scripts to ts
