+++ .gitignore
@@ -2,3 +2,6 @@ node_modules
 *.log
 .DS_Store
 dist
+src/*.js
+src/*.json
+!src/package.json
