+++ .nvmrc
@@ -1,1 +1,1 @@
-12.10.0
+13.8.0
