+++ .prettierignore
@@ -1,1 +1,2 @@
 CHANGELOG.md
+dist
