+++ .vscode/settings.json
@@ -9,5 +9,6 @@
   "yaml.schemas": {
     "http://json-schema.org/draft-07/schema": "*.schema.yaml"
   },
-  "yaml.completion": true
+  "yaml.completion": true,
+  "typescript.tsdk": "node_modules/typescript/lib"
 }
