+++ src/global.d.ts
@@ -1,0 +1,3 @@
+declare module 'deepdash'
+
+declare module 'json-transforms'
