+++ src/package.json
@@ -1,0 +1,3 @@
+{
+  "type": "module"
+}
