+++ tsconfig.json
@@ -1,14 +1,14 @@
 {
+  "include": ["src/*.ts"],
   "compilerOptions": {
-    "declaration": true,
+    "module": "ESNext",
+    "target": "ES2020",
     "noImplicitReturns": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "strict": true,
     "resolveJsonModule": true,
     "esModuleInterop": true,
-    "outDir": "dist"
-  },
-  "include": ["index.ts"],
-  "exclude": ["node_modules", "dist/**/*"]
+    "moduleResolution": "Node"
+  }
 }
