+++ tsconfig.schemas.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": [
+    "src/validate-reference-files.ts",
+    "src/validate-schemas.ts",
+    "src/index.ts"
+  ]
+}
